% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_wiki_data}
\alias{api_wiki_data}
\title{api_wiki_data}
\usage{
api_wiki_data(current_article = "Smoltification")
}
\arguments{
\item{current_article}{The article that should return a .json.}
}
\value{
A .json file of the article
}
\description{
This function uses the dbpedia.org api to get a specific article and returns it as a .json file.
Note: Keep in mind that dbpedia.org does not contain every article on wikipedia. https://dbpedia.org/ is also
case sensitive, so "Helen_Abbey" must be written as "Helen_Abbey" and not "helen_abbey", "Helen_abbey" or any other
variant.
}
\examples{
api_wiki_data(current_article = "Helen_Abbey")
}
