% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{parse_data}
\alias{parse_data}
\title{parse_data}
\usage{
parse_data(api_data, lan = "en")
}
\arguments{
\item{api_data}{The output you want to parse}

\item{lan}{The language you want to print it out in. You can choose betweeen English ("en"), Polish ("pl") and Swedish ("sv").}
}
\value{
A list with related topics and the article abstract.
}
\description{
parses an output from api_wiki_data
}
\examples{
parse_data(api_wiki_data(current_article = "Helen_Abbey"))


}
