% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{women_in_statistics}
\alias{women_in_statistics}
\title{women_in_statistics}
\format{
\describe{
This is just a vector with subfolders. For example "Helen_Abbey" is a string that describes a page. So it refers to https://en.wikipedia.org/wiki/Helen_Abbey.

}
}
\source{
List of women in statistics—Wikipedia. Retrieved October 1, 2025, from \href{https://en.wikipedia.org/wiki/List_of_women_in_statistics}{Wikipedia}
}
\usage{
data(women_in_statistics)
}
\description{
Sample data for this package. This contains the links to all the statisticians mentioned in the list.
}
\keyword{datasets}
